UPDATE Transform int to double

+++ zooShell/src/main/java/doyenm/zooshell/model/SizeAttributes.java
@@ -11,10 +11,10 @@ import lombok.Setter;
 @Setter
 public class SizeAttributes {
 
-    private int femaleWeight;
-    private int maleWeight;
+    private double femaleWeight;
+    private double maleWeight;
 
-    public SizeAttributes(int femaleWeight, int maleWeight) {
+    public SizeAttributes(double femaleWeight, double maleWeight) {
         this.femaleWeight = femaleWeight;
         this.maleWeight = maleWeight;
     }
